MongoDBHelper class moved to 'uk.ac.ebi.eva.utils'
