+++ src/main/java/uk/ac/ebi/eva/pipeline/io/writers/GeneWriter.java
@@ -4,8 +4,8 @@ import org.opencb.datastore.core.ObjectMap;
 import org.springframework.batch.item.data.MongoItemWriter;
 import org.springframework.data.mongodb.core.MongoOperations;
 
-import uk.ac.ebi.eva.pipeline.MongoDBHelper;
 import uk.ac.ebi.eva.pipeline.model.FeatureCoordinates;
+import uk.ac.ebi.eva.utils.MongoDBHelper;
 
 public class GeneWriter extends MongoItemWriter<FeatureCoordinates> {
